requirements.txt
--index-url https://mirrors.aliyun.com/pypi/simple

apscheduler==3.6.3
arrow==0.17.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
click==7.1.2; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
flask-cors==3.0.9
flask-marshmallow==0.14.0
flask-sqlalchemy==2.4.4; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
flask-wtf==0.14.3
flask==1.1.2; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
itsdangerous==1.1.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
jinja2==2.11.2; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
markupsafe==1.1.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
marshmallow-sqlalchemy==0.24.0; python_version >= "3.6"
marshmallow==3.9.0; python_version >= "3.6"
numpy==1.19.3; python_version >= "3.6" and python_full_version >= "3.6.1"
pandas==1.1.4; python_full_version >= "3.6.1"
pymysql==0.10.1
pyodbc==4.0.30
python-dateutil==2.8.1; python_full_version >= "3.6.1"
python-dotenv==0.15.0
pytz==2020.1; python_full_version >= "3.6.1"
six==1.15.0; python_full_version >= "3.6.1"
sqlalchemy-utils==0.36.8
sqlalchemy==1.3.20; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
tzlocal==2.1
werkzeug==1.0.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
wtforms==2.3.3
